refactor: expose report preview on ADR
Make adr.preview_report(...) the public API while keeping the WSGI runner and preview orchestration private in serverless.preview. Remove the standalone package export and align tests and documentation with the method-based interface.

Verified with focused preview and ADR unit tests, make smoketest, make check with the Windows-blocked license hook skipped, and a documentation build.
