+++ src/ansys/dynamicreporting/core/serverless/__init__.py
@@ -58,11 +58,9 @@ from .template import (
     TreeMergeGenerator,
     UserDefinedLayout,
 )
-from .preview import preview_report
 
 __all__ = [
     "ADR",
-    "preview_report",
     "Session",
     "Dataset",
     "Item",
